+++ src/View/View/AppView.php
@@ -140,6 +140,14 @@ class AppView extends BaseView {
 		//Assigns the title for layout
 		$this->assign('title', $this->_getTitleForLayout());
 		
+		//Adds the favicon
+		if(is_readable(WWW_ROOT.'favicon.ico'))
+			$this->Html->meta('icon');
+		
+		//Adds the "theme color" (the toolbar color for some mobile browser)
+		if(config('frontend.toolbar_color'))
+			$this->Html->meta('theme-color', config('frontend.toolbar_color'));
+		
 		//Adds the meta tag for RSS posts
 		if(config('frontend.rss_meta'))
 			$this->Html->meta(__d('me_cms', 'Latest posts'), '/posts/rss', ['type' => 'rss']);